extracted onAppointmentsFormSubmit to separate file, NIT: changed name of utils to formValidation utils for better semantics, removed comments from index.js
